Move from debug server to beacon server
