+++ beacon-chain/rpc/beacon/validators.go
@@ -12,6 +12,7 @@ import (
 	"github.com/prysmaticlabs/prysm/beacon-chain/core/state"
 	"github.com/prysmaticlabs/prysm/beacon-chain/core/validators"
 	"github.com/prysmaticlabs/prysm/beacon-chain/flags"
+	pbrpc "github.com/prysmaticlabs/prysm/proto/beacon/rpc/v1"
 	"github.com/prysmaticlabs/prysm/shared/bytesutil"
 	"github.com/prysmaticlabs/prysm/shared/featureconfig"
 	"github.com/prysmaticlabs/prysm/shared/pagination"
@@ -1044,6 +1045,92 @@ func (bs *Server) GetValidatorPerformance(
 		BalancesBeforeEpochTransition: beforeTransitionBalances,
 		BalancesAfterEpochTransition:  afterTransitionBalances,
 		MissingValidators:             missingValidators,
+	}, nil
+}
+
+// GetIndividualVotes retrieves individual voting status of validators.
+func (ds *Server) GetIndividualVotes(
+	ctx context.Context,
+	req *pbrpc.IndividualVotesRequest,
+) (*pbrpc.IndividualVotesRespond, error) {
+	if !featureconfig.Get().NewStateMgmt {
+		return nil, status.Error(codes.FailedPrecondition, "Requires --enable-new-state-mgmt to function")
+	}
+
+	currentEpoch := helpers.SlotToEpoch(ds.GenesisTimeFetcher.CurrentSlot())
+	if req.Epoch > currentEpoch {
+		return nil, status.Errorf(
+			codes.InvalidArgument,
+			"Cannot retrieve information about an epoch in the future, current epoch %d, requesting %d",
+			currentEpoch,
+			req.Epoch,
+		)
+	}
+
+	requestedState, err := ds.StateGen.StateBySlot(ctx, helpers.StartSlot(req.Epoch))
+	if err != nil {
+		return nil, status.Errorf(codes.Internal, "Could not retrieve archived state for epoch %d: %v", req.Epoch, err)
+	}
+	// Track filtered validators to prevent duplication in the response.
+	filtered := map[uint64]bool{}
+	filteredIndices := make([]uint64, 0)
+	votes := make([]*pbrpc.IndividualVotesRespond_IndividualVote, 0, len(req.Indices)+len(req.PublicKeys))
+	// Filter out assignments by public keys.
+	for _, pubKey := range req.PublicKeys {
+		index, ok := requestedState.ValidatorIndexByPubkey(bytesutil.ToBytes48(pubKey))
+		if !ok {
+			votes = append(votes, &pbrpc.IndividualVotesRespond_IndividualVote{PublicKey: pubKey, ValidatorIndex: ^uint64(0)})
+			continue
+		}
+		filtered[index] = true
+		filteredIndices = append(filteredIndices, index)
+	}
+	// Filter out assignments by validator indices.
+	for _, index := range req.Indices {
+		if !filtered[index] {
+			filteredIndices = append(filteredIndices, index)
+		}
+	}
+	sort.Slice(filteredIndices, func(i, j int) bool {
+		return filteredIndices[i] < filteredIndices[j]
+	})
+
+	v, b, err := precompute.New(ctx, requestedState)
+	if err != nil {
+		return nil, status.Errorf(codes.Internal, "Could not set up pre compute instance: %v", err)
+	}
+	v, b, err = precompute.ProcessAttestations(ctx, requestedState, v, b)
+	if err != nil {
+		return nil, status.Error(codes.Internal, "Could not pre compute attestations")
+	}
+	validators := requestedState.ValidatorsReadOnly()
+	for _, index := range filteredIndices {
+		if int(index) >= len(v) {
+			votes = append(votes, &pbrpc.IndividualVotesRespond_IndividualVote{ValidatorIndex: index})
+			continue
+		}
+		pb := validators[index].PublicKey()
+		votes = append(votes, &pbrpc.IndividualVotesRespond_IndividualVote{
+			Epoch:                            req.Epoch,
+			PublicKey:                        pb[:],
+			ValidatorIndex:                   index,
+			IsSlashed:                        v[index].IsSlashed,
+			IsWithdrawableInCurrentEpoch:     v[index].IsWithdrawableCurrentEpoch,
+			IsActiveInCurrentEpoch:           v[index].IsActiveCurrentEpoch,
+			IsActiveInPreviousEpoch:          v[index].IsActivePrevEpoch,
+			IsCurrentEpochAttester:           v[index].IsCurrentEpochAttester,
+			IsCurrentEpochTargetAttester:     v[index].IsCurrentEpochTargetAttester,
+			IsPreviousEpochAttester:          v[index].IsPrevEpochAttester,
+			IsPreviousEpochTargetAttester:    v[index].IsPrevEpochTargetAttester,
+			IsPreviousEpochHeadAttester:      v[index].IsPrevEpochHeadAttester,
+			CurrentEpochEffectiveBalanceGwei: v[index].CurrentEpochEffectiveBalance,
+			InclusionSlot:                    v[index].InclusionSlot,
+			InclusionDistance:                v[index].InclusionDistance,
+		})
+	}
+
+	return &pbrpc.IndividualVotesRespond{
+		IndividualVotes: votes,
 	}, nil
 }
